Metafile changes
Uses dummy_mod_frag_model instead of empty models.
Adds a bunch of strings as the modShopLabel.
Fixes wheelScale, once and for all.
Makes the dmc_debug use a placeholder string.
Fixes the LOD Distance for the DMC-12. Should improve performance, may help with world unloading.
Adds a 'standard' for the timeStamp on setup2.xml. Please change it whenever you make changes to the dlc.rpf, thank you in advance.